Fix bug for lazy loading single associations
When a single association (i.e. not a list) is loaded, @object not a Flexirest::ResultIterator, but a Flexirest::Base object. This object does not behave like an array, but as a hash. I needed to add a fix to yield the parameters as if it is a hash in case @object is a Flexirest::Base object.