Fix fetching included nested associations in json:api proxy
I experienced some problems with fetching nested associations from the compound document (= included key from a response). They were regularly not included in the response. I fixed this issue, and I also added the  or  parameters to the parsed response object which are filled with the IDs of the associated resources. Furthermore, I add LazyAssociationsLoaders in case a record cannot be found in the compound document, instead of the resource yielding nil whereas the associated ID may be non-nil.