Update usage of Faraday error classes
Faraday 1.0.0 got out and updated the hierarchy of its error classes:
https://github.com/lostisland/faraday/blob/ff9dc1d1219a1bbdba95a9a4cf5d135b97247ee2/UPGRADING.md

We must now use `Faraday::TimeoutError` instead of
`Faraday::Error::TimeoutError`.

It is worth noting that using `Faraday::TimeoutError` works with
the older 0.17.3 version of Faraday, so this change will not
impact projects that do not wish to upgrade to Faraday 1.0
(I successfully ran rspec locally with both Faraday versions).