Prevent crash on JSONAPI invalid error response
According to the spec:

    The members data and errors MUST NOT coexist in the same document.

(https://jsonapi.org/format/#document-top-level)

However sometimes a server will produce an error response which include both
a (legit) `"errors"` and an (invalid) `"data"` key, which so far
caused Flexirest to crash instead of raising the proper Flexirest exception.

This fix prevents the crash and makes Flexirest raise the relevant exception.